requirements.txt
attrs==21.4.0
Automat==20.2.0
beautifulsoup4==4.10.0
cffi==1.15.0
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
greenlet==1.1.2
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.8.0
mysql-connector-python==8.0.28
numpy==1.22.2
packaging==21.3
pandas==1.4.1
parsel==1.6.0
Pillow==9.0.1
priority==1.3.0
Protego==0.2.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytesseract==0.3.9
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
queuelib==1.6.2
scraper-helper==0.0.8
Scrapy==2.5.1
scrapy-user-agents==0.1.1
Scrapy-UserAgents==0.0.1
service-identity==21.1.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typing_extensions==4.1.1
ua-parser==0.10.0
urllib3==1.26.8
user-agents==2.2.0
w3lib==1.22.0
zope.interface==5.4.0
